Public support for template
