+++ template-skeleton-cn/template/src/routes.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import { Navigate } from 'react-router-dom';
-import { OAuthCallback } from '@carto/react-auth';
 import Main from 'components/views/Main';
 import NotFound from 'components/views/NotFound';
 // [hygen] Import views
@@ -14,7 +13,6 @@ const routes = [
       // [hygen] Add routes
     ],
   },
-  { path: '/oauthCallback', element: <OAuthCallback /> },
   { path: '404', element: <NotFound /> },
   { path: '*', element: <Navigate to='/404' /> },
 ];
